a mettre la norme

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mzhan <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/17 15:09:25 by mzhan             #+#    #+#             */
-/*   Updated: 2021/04/06 14:57:23 by mzhan            ###   ########.fr       */
+/*   Updated: 2021/04/07 16:45:00 by mzhan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ int ft_printf(const char *format, ...);
 void ft_convert_percentage_s(va_list *arguments, t_struct *flags);
 void ft_convert_percentage_d(va_list *arguments, t_struct *flags);
 void ft_convert_percentage_u(va_list *arguments, t_struct *flags);
-int ft_nbrlen(int res);
+int ft_nbrlen(long int res);
 void ft_putunsign_fd(long unsigned int n, int fd);
 void ft_convert_percentage_x(va_list *arguments, t_struct *flags);
 void ft_putnbr_hexa(long unsigned int nbr, char *base);
